+++ src/MainTransformer/MainTransformerInterface.php
@@ -19,12 +19,11 @@ interface MainTransformerInterface
 {
     /**
      * @param array<array-key,Type> $targetTypes If provided, it will be used instead of guessing the type
-     * @param array<string,mixed> $context
      */
     public function transform(
         mixed $source,
         mixed $target,
         array $targetTypes,
-        array $context
+        Context $context,
     ): mixed;
 }