+++ src/Mapper.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Rekalogika\Mapper;
 
 use Rekalogika\Mapper\Exception\UnexpectedValueException;
+use Rekalogika\Mapper\MainTransformer\Context;
 use Rekalogika\Mapper\MainTransformer\MainTransformerInterface;
 use Rekalogika\Mapper\Util\TypeFactory;
 
@@ -52,7 +53,7 @@ final class Mapper implements MapperInterface
             source: $source,
             target: $target,
             targetTypes: [$targetType],
-            context: []
+            context: Context::create(),
         );
 
         if ($target === null) {