+++ src/Transformer/Contracts/TransformerInterface.php
@@ -13,30 +13,19 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Transformer\Contracts;
 
-use Rekalogika\Mapper\Exception\ExceptionInterface;
-use Rekalogika\Mapper\Exception\InvalidArgumentException;
-use Rekalogika\Mapper\Exception\LogicException;
-use Rekalogika\Mapper\ObjectCache\Exception\CircularReferenceException;
+use Rekalogika\Mapper\MainTransformer\Context;
 use Symfony\Component\PropertyInfo\Type;
 
 interface TransformerInterface
 {
     public const MIXED = 'mixed';
 
-    /**
-     * @param array<string,mixed> $context
-     *
-     * @throws InvalidArgumentException
-     * @throws CircularReferenceException
-     * @throws LogicException
-     * @throws ExceptionInterface
-     */
     public function transform(
         mixed $source,
         mixed $target,
         ?Type $sourceType,
         ?Type $targetType,
-        array $context
+        Context $context,
     ): mixed;
 
     /**