+++ src/Transformer/TraversableToArrayAccessTransformer.php
@@ -16,7 +16,8 @@ namespace Rekalogika\Mapper\Transformer;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Rekalogika\Mapper\Exception\InvalidArgumentException;
-use Rekalogika\Mapper\MainTransformer\MainTransformer;
+use Rekalogika\Mapper\MainTransformer\Context;
+use Rekalogika\Mapper\ObjectCache\ObjectCache;
 use Rekalogika\Mapper\Transformer\Contracts\MainTransformerAwareInterface;
 use Rekalogika\Mapper\Transformer\Contracts\MainTransformerAwareTrait;
 use Rekalogika\Mapper\Transformer\Contracts\TransformerInterface;
@@ -37,7 +38,7 @@ final class TraversableToArrayAccessTransformer implements TransformerInterface,
         mixed $target,
         ?Type $sourceType,
         ?Type $targetType,
-        array $context
+        Context $context
     ): mixed {
         if ($targetType === null) {
             throw new InvalidArgumentException('Target type must not be null.');
@@ -45,7 +46,7 @@ final class TraversableToArrayAccessTransformer implements TransformerInterface,
 
         // get object cache
 
-        $objectCache = MainTransformer::getObjectCache($context);
+        $objectCache = $context->get(ObjectCache::class);
 
         // The source must be a Traversable or an array (a.k.a. iterable).